v02/app/requirements.txt
Flask
mysql-connector-python
docker
